Fix restoring object proxies and view providers from old files
Increase version number
