--- Init.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#  Init.py
-#
-#  Copyright 2015 Shai Seger <shaise at gmail dot com>
-#
-#  This program is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 2 of the License, or (at your option) any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU Lesser General Public
-#  License along with this program; if not, write to the Free Software
-#  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-#  MA 02110-1301, USA.
-#
-#
-##############################################################################
-
-import SheetMetalBaseShape
-import SheetMetalBendWall
-import SheetMetalBaseBend
-import SheetMetalSolidBend
-import SheetMetalFoldWall
-import SheetMetalForming
-import SheetMetalJunction
-import SheetMetalRelief
-import SketchOnSheetMetal
-import SheetMetalCornerRelief
-import SheetMetalExtend
-import SheetMetalUnfolder
