+++ package.xml
@@ -2,7 +2,7 @@
 <package format="1" xmlns="https://wiki.freecad.org/Package_Metadata">
   <name>SheetMetal Workbench</name>
   <description>A simple sheet metal tools workbench for FreeCAD.</description>
-  <version>0.5.0</version>
+  <version>0.5.1</version>
   <date>2024-05-13</date>
   <maintainer email="shaise@gmail.com">Shai Seger</maintainer>
   <license file="LICENSE">LGPL-2.1-or-later</license>
